fix small edge-case bug in prefix filter of label intersection FJ
A prefix filter for checking set overlaps requires a total order. Since
when sorting by the inverted frequency there is the possibility that two
different elements have the same frequency, this only gives us a total
preorder when sorting by the frequency. In order to have a total order,
we also subsort by (label, type)-tuples which gives us a total
lexicographic order.